Implement access logs parser
