+++ homework11/README.md
@@ -43,3 +43,84 @@ root: 116
 Больше всего памяти использует: com.docker.hyperkit
 Больше всего CPU использует: softwareupdated
 ```
+
+
+Access Log parser
+===
+
+Цель
+---
+Написать скрипт анализа приложенного access.log файла
+Требования к реализации
+
+Должна быть возможность указать директорию где искать логи или конкретный файл
+
+Должна быть возможность обработки всех логов внутри одной дирректории
+
+Для access.log должна собираться следующая информация:
+
+общее количество выполненных запросов
+количество запросов по типу: GET - 20, POST - 10 и т.п.
+топ 3 IP адресов, с которых были сделаны запросы
+топ 3 самых долгих запросов, должно быть видно метод, url, ip, время запроса
+Собранная статистика должна быть сохранена в json файл и выведена в терминал в свободном (но понятном!) формате
+Должен быть README.md файл, который описывает как работает скрипт
+
+
+
+Как работает скрипт:
+---
+Скрипт принимает в качестве аргументов директорию для поиска лог файлов, прямой путь к файлу, парамерт --save для сохранения результатов сканирования.
+Приоритет имеет прямой путь к файлу.
+
+Более экономично было бы открывать файлы напрямую питоном, ноо в целях тренировки был выбран способ совмещения выполнения команды `cat` с последующим парсингов результата.
+
+Анализ результата парсинга происходит следующим образом:
+- self._collect_total_request_count() -> смотрим количество запросов
+- self._collect_method_stats() -> собираем статистику методов встроенным классом Counter
+- self._collect_ip_stats() -> собираем статистику для IP адрессов
+- self._collect_long_requests() -> собираем статистику для длятельности запросов (%D - микросекунды, потому далее делим на 1000000 для получания секунд)
+- генерируем вывод
+- опционально сохраняем JSON
+
+Что можно улучщить:
+- открывать через open("access.log") -> этим уменьшим расход памяти
+- создавать несколько потоков, для парсинга файлов частями -> этим ускорим обработку
+- добавить обработку несуществующих директорий и файлов
+
+
+Пример использования
+---
+```
+(.venv) ➜  Homework git:(redilonka/fea/access_parser) ✗ python homework11/log_parser.py -d /Users/redilonka/code        
+Following command is about to start: cat /Users/redilonka/code/access.log
+Общее количество выполненных запросов: 3207700
+
+Количество запросов по типу: GET: 2238870, POST: 918776, HEAD: 49968, PUT: 70, OPTIONS: 15, DELETE: 1
+
+Топ 3 IP адресов, с которых были сделаны запросы: 
+IP: 193.106.31.130: 235209 запросов
+IP: 198.50.156.189: 167812 запросов
+IP: 5.112.235.245: 166722 запросов
+
+Топ 3 самых долгих запросов: 
+URL: /administrator/index.php IP: 62.93.172.245 Time: 0.01 сек
+URL: /media/system/css/modal.css IP: 213.162.68.113 Time: 0.01 сек
+URL: /images/phocagallery/almhuette/thumbs/phoca_thumb_m_terasse.jpg IP: 194.166.55.43 Time: 0.01 сек
+
+(.venv) ➜  Homework git:(redilonka/fea/access_parser) ✗ python homework11/log_parser.py -f ../access.log           
+Following command is about to start: cat ../access.log
+Общее количество выполненных запросов: 3207700
+
+Количество запросов по типу: GET: 2238870, POST: 918776, HEAD: 49968, PUT: 70, OPTIONS: 15, DELETE: 1
+
+Топ 3 IP адресов, с которых были сделаны запросы: 
+IP: 193.106.31.130: 235209 запросов
+IP: 198.50.156.189: 167812 запросов
+IP: 5.112.235.245: 166722 запросов
+
+Топ 3 самых долгих запросов: 
+URL: /administrator/index.php IP: 62.93.172.245 Time: 0.01 сек
+URL: /media/system/css/modal.css IP: 213.162.68.113 Time: 0.01 сек
+URL: /images/phocagallery/almhuette/thumbs/phoca_thumb_m_terasse.jpg IP: 194.166.55.43 Time: 0.01 сек
+```
